backend/requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
scikit-learn==1.6.1
joblib==1.4.2
pandas==2.2.3
numpy==2.2.3
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.10.0
python-whois==0.9.5
pyzbar==0.1.9
Pillow==11.1.0
python-multipart==0.0.20
